Refactor: 프로필 설정 화면 분리 #129
